update : move friends and userlist to follow folder
